backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.2
email-validator==2.2.0
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
SQLAlchemy==2.0.39
psycopg[binary]==3.2.6
alembic==1.15.2
aio-pika==9.4.3
redis==5.2.1
slowapi==0.1.9
boto3==1.37.15
